rename components to match the standard
